✨ Feat:#220-My challenge detail client 파일 분리
