+++ src/app/challengeList/page.tsx
@@ -1,6 +1,9 @@
 import { HydrationBoundary, QueryClient, dehydrate } from '@tanstack/react-query';
-import { fetchAdminChallenge, fetchChallenge, fetchRanker } from '@/api/challengeService';
+import Image from 'next/image';
+import { fetchAdminChallenge, fetchChallenge } from '@/api/challengeService';
 import ChallengeListClient from '@/components/ClientWrapper/ChallengeListClient';
+
+const S3_BASE_URL = process.env.NEXT_PUBLIC_S3_BASE_URL;
 
 export default async function ChallengeListPage() {
   const rankerData = [
@@ -39,8 +42,6 @@ export default async function ChallengeListPage() {
   });
   const dehydratedState = dehydrate(queryClient);
 
-  // const rankerData = await fetchRanker();
-
   const Data = {
     adminchallengeData,
     rankerData
@@ -52,7 +53,9 @@ export default async function ChallengeListPage() {
         {Object.values(Data).every(value => !!value) ? (
           <ChallengeListClient adminchallengeData={adminchallengeData} rankerData={rankerData} />
         ) : (
-          <p>Loading...</p>
+          <div className="flex w-full justify-center items-center min-h-screen">
+            <Image src={`${S3_BASE_URL}/loading.svg`} alt="loading" width={200} height={200} />
+          </div>
         )}
       </div>
     </HydrationBoundary>
